+++ marvel-api/src/main/java/com/gabriel/scavassa/marvelapi/repository/CharacterRepository.java
@@ -1,10 +1,12 @@
 package com.gabriel.scavassa.marvelapi.repository;
 
 import com.gabriel.scavassa.marvelapi.domain.Character;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CharacterRepository extends JpaRepository<Character, Long> {
+public interface CharacterRepository extends JpaRepository<Character, Integer> {
 
 }
