+++ marvel-api/src/main/java/com/gabriel/scavassa/marvelapi/repository/CreatorsRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CreatorsRepository extends JpaRepository<Creators, Long> {
+public interface CreatorsRepository extends JpaRepository<Creators, Integer> {
 
 }
