+++ marvel-api/src/main/java/com/gabriel/scavassa/marvelapi/repository/comic/ComicDateRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ComicDateRepository extends JpaRepository<ComicDate, Long> {
+public interface ComicDateRepository extends JpaRepository<ComicDate, Integer> {
 
 }
